+++ demo/index.html
@@ -4,8 +4,8 @@
         <meta charset="UTF-8">
         <meta name="viewport" content="width=device-width, initial-scale=1.0">
         <title>Seven Segment Display Reader</title>
-        <link rel="stylesheet" href="styles.css">
-        <script src="index.js" type="module"></script>
+        <link rel="stylesheet" href="./styles.css">
+        <script src="./index.js" type="module"></script>
     </head>
     <body>
         <!-- Config button in top right -->
