+++ demo/styles.css
@@ -1,0 +1,173 @@
+body {
+    font-family: Arial, sans-serif;
+    text-align: center;
+    position: relative;
+}
+
+button {
+    margin: 10px;
+    padding: 10px 20px;
+    font-size: 18px;
+}
+
+h1 {
+    margin-top: 8px;
+    border-bottom: 5px double;
+}
+
+*.hidden {
+    display: none !important;
+}
+
+#video-wrapper {
+    position: relative;
+    float: left;
+
+    canvas {
+        display: block;
+        margin: 0 15px;
+    }
+
+    video {
+        display: block;
+        margin: 0 15px;
+    }
+
+    video.rotated {
+        transform: rotate(180deg);
+    }
+}
+
+#history {
+    height: 512px;
+    overflow-y: scroll;
+    border: 1px solid #555555;
+    margin: 10px;
+    text-align: left;
+}
+
+#message {
+    font-size: 18px;
+    margin: 10px;
+    text-align: left;
+}
+
+#output {
+    font-size: 36px;
+    font-weight: bold;
+    font-family: system-ui;
+    letter-spacing: 1px;
+}
+
+/* Config button styles */
+.config-button {
+    position: fixed;
+    top: 10px;
+    right: 10px;
+    width: 40px;
+    height: 40px;
+    padding: 0;
+    font-size: 24px;
+    background: #f0f0f0;
+    border: 1px solid #ccc;
+    border-radius: 50%;
+    cursor: pointer;
+    z-index: 1000;
+    transition: background-color 0.3s;
+}
+
+.config-button:hover {
+    background: #e0e0e0;
+}
+
+/* Config slideover panel styles */
+.config-slideover {
+    position: fixed;
+    top: 0;
+    right: -350px;
+    width: 350px;
+    height: 100%;
+    background: white;
+    box-shadow: -2px 0 10px rgba(0, 0, 0, 0.1);
+    transition: right 0.3s ease;
+    z-index: 1001;
+}
+
+.config-slideover.open {
+    right: 0;
+}
+
+.config-slideover-content {
+    height: 100%;
+    display: flex;
+    flex-direction: column;
+}
+
+.config-header {
+    display: flex;
+    justify-content: space-between;
+    align-items: center;
+    padding: 20px;
+    border-bottom: 1px solid #e0e0e0;
+    background: #f8f8f8;
+}
+
+.config-header h2 {
+    margin: 0;
+    font-size: 20px;
+}
+
+.close-config {
+    background: none;
+    border: none;
+    font-size: 32px;
+    cursor: pointer;
+    padding: 0;
+    width: 30px;
+    height: 30px;
+    display: flex;
+    align-items: center;
+    justify-content: center;
+}
+
+.close-config:hover {
+    background: #e0e0e0;
+    border-radius: 50%;
+}
+
+.config-body {
+    padding: 20px;
+    overflow-y: auto;
+    flex: 1;
+}
+
+.config-body button {
+    width: 100%;
+    margin: 10px 0;
+    text-align: left;
+}
+
+.config-item {
+    margin: 15px 0;
+    padding: 10px;
+    background: #f5f5f5;
+    border-radius: 5px;
+}
+
+.config-item label {
+    display: flex;
+    align-items: center;
+    cursor: pointer;
+    font-size: 14px;
+}
+
+.config-item input[type="checkbox"] {
+    margin-right: 10px;
+    width: 18px;
+    height: 18px;
+    cursor: pointer;
+}
+
+.config-item span {
+    flex: 1;
+}
