+++ src/classes/SegmentDisplayReaderConfiguration.js
