+++ src/timingWorker.js
